requirements.txt
streamlit
opencv-python
numpy
Pillow
torch
torchvision
playsound


streamlit>=1.18
opencv-python
numpy
